+++ packages/extensions/modelerfour/test/modeler/modelerfour.requests.test.ts
@@ -2,188 +2,13 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 
 import assert from "assert";
-import { CodeModel, DictionarySchema, HttpHeader, Operation, Parameter, SealedChoiceSchema } from "@autorest/codemodel";
-import { HttpOperation, JsonType, ParameterLocation, RequestBody } from "@azure-tools/openapi";
+import { CodeModel, DictionarySchema, HttpHeader, Operation, Parameter } from "@autorest/codemodel";
+import { JsonType, ParameterLocation } from "@azure-tools/openapi";
 import * as oai3 from "@azure-tools/openapi";
 import { addOperation, createTestSpec, findByName } from "../utils";
-import { runModeler } from "./modelerfour-utils";
-
-async function runModelerWithOperation(
-  method: string,
-  path: string,
-  operation: oai3.HttpOperation,
-): Promise<Operation> {
-  const spec = createTestSpec();
-
-  addOperation(spec, path, {
-    [method]: operation,
-  });
-
-  const codeModel = await runModeler(spec);
-  const m4Operation = codeModel.operationGroups[0]?.operations[0];
-  assert(m4Operation);
-  return m4Operation;
-}
+import { runModeler, runModelerWithOperation } from "./modelerfour-utils";
 
 describe("Modelerfour.Request", () => {
-  describe("Body", () => {
-    const runModelerWithBody = async (body: RequestBody): Promise<Operation> => {
-      return runModelerWithOperation("post", "/test", {
-        requestBody: {
-          ...body,
-        },
-        responses: {},
-      });
-    };
-
-    describe("Required attribute", () => {
-      const runModelerWithBodyAndGetParam = async (body: RequestBody) => {
-        const operation = await runModelerWithBody(body);
-        const parameter = operation.requests?.[0]?.parameters?.[0];
-
-        expect(parameter).not.toBeNull();
-        return parameter;
-      };
-
-      const defaultBody = {
-        content: {
-          "application/octet-stream": {
-            schema: {
-              type: JsonType.Object,
-              format: "file",
-            },
-          },
-        },
-      };
-
-      it("mark body as required if required: true", async () => {
-        const parameter = await runModelerWithBodyAndGetParam({ ...defaultBody, required: true });
-        expect(parameter?.required).toBe(true);
-      });
-
-      it("mark body as not required if required: false", async () => {
-        const parameter = await runModelerWithBodyAndGetParam({ ...defaultBody, required: true });
-        expect(parameter?.required).toBe(true);
-      });
-
-      it("mark body as not required by default", async () => {
-        const parameter = await runModelerWithBodyAndGetParam(defaultBody);
-        expect(parameter?.required).toBe(undefined);
-      });
-    });
-
-    describe("multiple binary content-types", () => {
-      let operation: Operation;
-
-      beforeEach(async () => {
-        operation = await runModelerWithBody({
-          content: {
-            "application/octet-stream": {
-              schema: { type: JsonType.String, format: "binary" },
-            },
-            "image/png": {
-              schema: { type: JsonType.String, format: "binary" },
-            },
-          },
-        });
-      });
-
-      it("only create one request", async () => {
-        expect(operation.requests?.length).toEqual(1);
-      });
-
-      it("operation has a content-type parameter", async () => {
-        const param = findByName("content-type", operation.requests?.[0].parameters);
-        assert(param);
-        expect(param?.schema.type).toEqual("sealed-choice");
-        const schema = param?.schema as SealedChoiceSchema;
-        expect(schema.choices.map((x) => x.value)).toEqual(["application/octet-stream", "image/png"]);
-      });
-    });
-
-    describe("has binary content type and application/json with binary schema", () => {
-      let operation: Operation;
-
-      beforeEach(async () => {
-        operation = await runModelerWithBody({
-          content: {
-            "application/octet-stream": {
-              schema: { type: JsonType.String, format: "binary" },
-            },
-            "application/json": {
-              schema: { type: JsonType.String, format: "binary" },
-            },
-          },
-        });
-      });
-
-      it("only create one request", async () => {
-        expect(operation.requests?.length).toEqual(1);
-      });
-
-      it("operation has a content-type parameter", async () => {
-        const param = findByName("content-type", operation.requests?.[0].parameters);
-        assert(param);
-        expect(param?.schema.type).toEqual("sealed-choice");
-        const schema = param?.schema as SealedChoiceSchema;
-        expect(schema.choices.map((x) => x.value)).toEqual(["application/json", "application/octet-stream"]);
-      });
-    });
-  });
-
-  describe("multipart/form-data", () => {
-    let parameters: Parameter[] | undefined;
-
-    beforeEach(async () => {
-      const spec = createTestSpec();
-      const operation: HttpOperation = {
-        requestBody: {
-          content: {
-            "multipart/form-data": {
-              schema: {
-                type: JsonType.Object,
-                properties: {
-                  id: {
-                    type: JsonType.String,
-                  },
-                  address: {
-                    type: JsonType.String,
-                  },
-                },
-              },
-            },
-          },
-        },
-        parameters: [{ in: ParameterLocation.Query, name: "queryParam", schema: { type: JsonType.String } }],
-        responses: { "200": { description: "Ok." } },
-      };
-
-      addOperation(spec, "/test", {
-        post: operation,
-      });
-
-      const codeModel = await runModeler(spec);
-      parameters = codeModel.operationGroups[0]?.operations[0]?.requests?.[0]?.parameters;
-      assert(parameters);
-    });
-
-    it("mark body parameter as isInMultipart", async () => {
-      const idParameter = parameters?.[0];
-      const addressParameter = parameters?.[1];
-      expect(idParameter?.language.default.name).toEqual("id");
-      expect(idParameter?.isPartialBody).toBe(true);
-      expect(addressParameter?.language.default.name).toEqual("address");
-      expect(idParameter?.isPartialBody).toBe(true);
-    });
-
-    it("doesn't mark other parameter as isInMultipart", async () => {
-      const queryParam = parameters?.[2];
-
-      expect(queryParam?.language.default.name).toEqual(queryParam);
-      expect(queryParam?.isPartialBody).toBeFalsy();
-    });
-  });
-
   describe("x-ms-header-collection-prefix headers", () => {
     let spec: oai3.Model;
